form: Refactor to utilise the form.html template
This refactor utilises the form.html template for refactoring, rather than using the separate templates for each page that inherit from form.html. This also means that a title block has been added to form.html to ensure the correct title is shown for the action that has been chosen by the user.