validation: Ask user if they want to remove an object
Currently, a superuser can remove objects from a model without needing to validate their choice. This could lead to mistakes and deletion of incorrect objects. There is now a page where the user will be asked if they want to remove an object, and only when they press delete again, will it be deleted.